Only send email notification if there was a change since last time
In the process of setting up test data, noticed an error that
RoleActivity was not being spotted properly, because
RoleActivity--Role--Organization, not RoleActivity--Organization.

Tests FTW.

Also updated caching logic to show data as at the latest cached time.
This is helpful if looking at data the day after an update but before the
next update. The stats won't have changed but the stats were no longer
last updated 'today'