Merge sameAsHigh
Profiling the `merge_query` with index was marginally faster than without index, down to NodeByLabelScan in first step being replaced by NodeByIndexScan.

It's only Organizations that have sameAsHigh, but the internal doc is is at the Resource level, so setting the index at Resource level and running the query at Resource level to take advantage of the index and eke out this little performance gain.