Refactored merge sameAsHigh
Phew, that was a bit of a mission. Built some test cases that highlighted
some things that had been missed in then merging logic. It seems to
work well now, but there is one known limitation.

A sameAs B
B is merged into A and B is deleted

Later

B is recreated but this time without sameAs A
A still has B in its internalSameAsXXXList
and B wouldn't be attached to A.

Not sure if this is real problem, let's see what the data shows over the
coming weeks.