Implemented tracking for industry/geo combos
Agonized over whether to store industry topicId, uri or name in
TrackedIndustryGeo table. Eventually decided that topicId is too
meaningless. Name (with comma separated words) is used throughout the UI
so eventually decided to stick with that.

However for geo I went with the geo code because that's a standard way
of referring to countries.regions