Implemented Family Tree
- Implement Family Tree page in place of parent-child pages
- Update same-as logic, leading to completely refactoring the organizational linkages page (`graph_utils.py`)

Phew bit of a mammoth commit here. Made the fatal error of thinking
"while I'm sorting out Feature X I may as well fix this other thing".

In this case "Feature X" is implementing a family tree graph showing parent-child relationships, to replace several earlier pages. It's a better UI and keeps things consistent.

The "this other thing" was the standard graph page.

Under the covers sorting out this feature meant a big change in how same-as relationships are handled between nodes. Previously they were being merged with `apoc.mergeNodes`. With this approach there was no way to disentangle a node that was merged incorrectly. Now relationships are copied but old nodes remain to aid in audit trail and debugging. This broke some other pages.

Fixing the standard graph page to handle this change ended up in re-writing `graph_utils.py` to be a lot more elegant. Also in the process the serialization for nodes has been made more consistent and prettier for displaying on the page. And realizing that there are more changes to implement around handling geonames nodes. But those changes can wait till later.

For now this commit is good enough.

Note to self: next time keep focussed on one feature in one commit only.