Fixed family tree SYR-23
This started off to specifically fix a bug where a node was being added into the family tree twice.

Once that was fixed and tested on live data I noticed another bug where the main node wouldn't be shown if there was no parent. So I fixed that too.

Then noticed a usability thing where children weren't listed in alphabetical order. This was a relatively quick fix so brought that into the scope of this ticket too.

Finally, there is a situation where orgs with same name can be repeated. That was a bigger piece of work so left that for another commit