Implemented "combine_same_as_name_only"
If this query_string param is set to 1 then all organizations with a "sameAsNameOnly" relationship are treated as the same entity and so are merged on the fly. They are not merged permanently the same way as "sameAsHigh"
relationships are.

Default is to have combine_same_as_name_only as 1 (on).