Seach results: use date rather than datetime when counting documents
The from date/datetime for counting how many documents over the past 2 years is cached. If the from date is a date then the cache will be hit each time we do a search on the same date. If the from date is a datetime then cache hit is very unlikely.