Changed organizationNames to orgName
organizationNames is an attribute on PartnershipActivity where you would expect more than one organization. So it makes some sense to be pluralized. However, many other attributes are also lists of items, even though they are have singular names. In the interests of consistency it makes more sense to have the field name singular