Added anon login account
I want to keep pages accessible to genuine web users, but not to be easily accessible to bots and scrapers. Let's see if this does the trick.

It's another good example of something that I thought would be pretty quick turned out to be a few hours work, even with ChatGPT's help. In this case, ChatGPT was helpful in some cases but a time-waster in others. It's a common pattern I'm seeing with the tech.