Created one model for all TrackedItems
This is now being set based on the industry/geo search feature

One side-effect is that if I search for industry "foo" and then select a region for all of the returned industries then the TrackedItem will think I want to track all activity in the region, not just the activity relating to "foo".
May need enhancing so user can differentiate between "I want to track everything in Turkey" vs "I want to track everything related to these industries in Turkey".