Only select industries for the chosen Geo
In the previous commit, if I choose one Geo it will select all industries within that Geo. This makes sense if the search is by Geo initially, but currently the search is by Industry first. So if selecting all industries for a Geo, make sure to list the matching industries.