Combine equivalent/similar activities
Often we get the same story repeated in a few distinct topics (e.g. in the title and the first line of the text).

Where there are two activities from the same Internal Document ID, if one has the same connections (or a subset of connections) of the other node then combine them. This will involve adding more weight to some relationships.