Turn off search for companies by industry text
An example of writing some code that
- didn't add a huge amount of value
- wasn't very performant

So would have been better not to write it in the first place. It's there now, so hiding it behind a flag so can turn it off if needed (and the tests don't need re-writing because they can set the flag and work as before).