Phew a much too big commit, but glad it's done.
Reworks the sameAsNameOnly logic. Now, instead of the upstream system deciding whether two Organizations are sameAsNameOnly, just doing that with an index search. It's much more performant than the older logic.

The testing ttl files have all been edited manually to use this kind of logic so the content may not be 100% correct. But it's good enough for test purposes.