Ensure trackeditem notifications use the latest cached stats
If an import job goes across midnight then the datestamp of the latest cached items could be different to the datestamp of the latest import job. Latest cache datestamp is always set to the end of today. Use the latest cached items datestamp for filtering but the latest import job for the email notification.