Implemented two-version cache.
When a new data load finishes, precalculation is done in the inactive cache version.
Once finished the active version is switched.

This means that there is always some cached data to serve to users, even if you're reading the old version until the new one is calculated. It's better than either no data or an error :D