Cache get_queryset for activities
Apparently get_queryset is invoked twice - once when collecting the data in the first place and once when doing the pagination. So caching the results to avoid having to recalculate everything.

In the process, changed the cache keys. The current version would remove [] so []_Foo_[]_Bar and []_Foo_Bar_[] would both become _Foo_, but both mean different things