Expanding date range for activities
If you don't provide a "days_ago" parameter and are looking for organization activities then it will start at 7 days and then increase the time-range if insufficient records are found.

No change to search by industry.

This was implemented because Claude Desktop would often do 2 API calls - one with a shorter timeframe and one with a longer timeframe. Simplified by doing it all in one go.