Merging unmerged resources
If new Node A is loaded which is sameAsHigh Node B and Node B has already been merged then Node A will never get merged.

This is different to a re-merge which is about cleaning up missing relationships to already-merged nodes.

Not yet integrated into the RDFPostProcessor steps, so would need to be run manually